64-bit filesize issue in squidclient if trying to post a file > 2GB

+++ tools/squidclient.cc
@@ -387,7 +387,7 @@ main(int argc, char *argv[])
             strcat(msg, buf);
         }
         if (put_fd > 0) {
-            snprintf(buf, BUFSIZ, "Content-length: %d\r\n", (int) sb.st_size);
+            snprintf(buf, BUFSIZ, "Content-length: %ld\r\n", (long int) sb.st_size);
             strcat(msg, buf);
         }
         if (opt_noaccept == 0) {
